+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+### v4.0.6 ( Aug 29, 2025 ) ###
+- **update:** Added translation support for dashboard analytics report scripts to improve internationalization (i18n).
+- **fix:** Resolved an issue in the AdminNotice Vue component by applying the Mixin utility for improved functionality.
+- **fix:** Added async search in store category in vendor edit.
+- **fix:** Admin not receiving withdrawal request email notifications.
+
+### v4.0.5 ( Jul 24, 2025 ) ###
+- **fix:** Improved script loading to ensure compatibility with WooCommerce versions above 10.0.2, preventing potential issues with script dependencies.
+
+### v4.0.4 ( Jul 17, 2025 ) ###
+
+- **update:** Admin Notice UI with Modern Design and Improved User Experience.
+- **update:** Enhanced Product Brand Integration with Dedicated Template and Manager Methods.
+- **update:** Enhanced product featured image selection with cropping functionality in the product editor.
+- **fix:** Vendor/store names with special characters (e.g., apostrophes) now display correctly in the vendor dropdown on the admin product edit screen.
+- **fix:** Display correct earning in vendor dashboard product add/edit page for different category.
+- **fix:** Update table header style for mobile display in order details page item list and withdraw approve, pending and cancel list.
+- **fix:** Improved how product inventory information is retrieved, ensuring more accurate and consistent display of stock values.
+- **fix:** Update table header style for mobile display in order details page item list and withdraw approve, pending and cancel list.
+- **fix:** Resolve an issue for "sold individually" option does not save on create or update a product.
+
+### v4.0.3 ( Jul 02, 2025 ) ###
+
+- **fix:** Add gradient background style for fa-threads social icon.
+- **fix** Resolved an issue where array access warnings in vendor balance calculation were being triggered.
+- **fix** Resolved an issue where incorrect timestamp on sale price schedule hampering product sales schedule added from vendor dashboard.
+- **fix** Resolved an issue where stock quantity always shows 0 in vendor dashboard product edit screen.
+- **fix** Improved the format of payment methods in the withdrawal settings to ensure they are consistently returned as a zero-based indexed list in the REST API response.
+
+
 ### v4.0.2 ( Jun 03, 2025 ) ###
 
 - **fix:** Resolved an issue where the single store page header style was broken on multiple themes.